[mlir][Interfaces] Move SubsetInsertionOpInterface to mlir/Interfaces
