+++ mlir/lib/Dialect/Tensor/Transforms/CMakeLists.txt
@@ -30,6 +30,7 @@ add_mlir_dialect_library(MLIRTensorTransforms
   MLIRMemRefDialect
   MLIRPass
   MLIRSCFDialect
+  MLIRSubsetInsertionOpInterface
   MLIRTensorDialect
   MLIRTensorUtils
   MLIRTilingInterface
