+++ mlir/lib/Interfaces/CMakeLists.txt
@@ -16,6 +16,7 @@ set(LLVM_OPTIONAL_SOURCES
   RuntimeVerifiableOpInterface.cpp
   ShapedOpInterfaces.cpp
   SideEffectInterfaces.cpp
+  SubsetInsertionOpInterface.cpp
   TilingInterface.cpp
   ValueBoundsOpInterface.cpp
   VectorInterfaces.cpp
@@ -82,6 +83,22 @@ add_mlir_interface_library(ParallelCombiningOpInterface)
 add_mlir_interface_library(RuntimeVerifiableOpInterface)
 add_mlir_interface_library(ShapedOpInterfaces)
 add_mlir_interface_library(SideEffectInterfaces)
+
+add_mlir_library(MLIRSubsetInsertionOpInterface
+  SubsetInsertionOpInterface.cpp
+
+  ADDITIONAL_HEADER_DIRS
+  ${MLIR_MAIN_INCLUDE_DIR}/mlir/Interfaces
+
+  DEPENDS
+  MLIRDestinationStyleOpInterface
+  MLIRSubsetInsertionOpInterfaceIncGen
+
+  LINK_LIBS PUBLIC
+  MLIRDestinationStyleOpInterface
+  MLIRIR
+  )
+
 add_mlir_interface_library(TilingInterface)
 add_mlir_interface_library(VectorInterfaces)
 add_mlir_interface_library(ViewLikeInterface)
